Separate SM4 algorithm from plugin
